fix: Drain the services error channel on restart.
This is a workaround that avoids the deadlock from the previous
implementation, as that would not exit the drain loop until the channel
was closed.

Here we just read if we can in a best effort fashion.

A better solution would be to let the writing goroutine close the
channel on exit and for us to notify on close with a context.