fix: Allow ibazel notifications to pass when not running tests (#52)
* fix: Handle all events in a single select.

If `ibazel` notifies of a change or the context is cancelled, then deal
with it when we are waiting for service errors or test completions.
This avoids a deadlock in interactive modes.

Fixes: #51

* refactor: Use a scoped context to cancel the test command.

* fix: Drain the services error channel on restart.

This is a workaround that avoids the deadlock from the previous
implementation, as that would not exit the drain loop until the channel
was closed.

Here we just read if we can in a best effort fashion.

A better solution would be to let the writing goroutine close the
channel on exit and for us to notify on close with a context.

* fix: Restore the `isOneShot` loop exit.

* ci: Run tests on PR.

---------

Co-authored-by: James-Wilkinson1_cba <james.wilkinson1@cba.com.au>